work on teacher model

+++ segmentation/train.py
@@ -279,7 +279,7 @@ def train_one_epoch(model, train_loader, device, optimizer, loss_function_train,
         # losses = loss_function_train(pred_scales, torch.nn.functional.one_hot(target_scales, num_classes=19))
         print(target_scales.max().item())
         losses = loss_function_train(pred_scales, torch.nn.functional.one_hot
-        (target_scales.long() - 1, num_classes=19).permute(0, 3, 1, 2))
+        (target_scales.long() - 1, num_classes=19).permute(0, 3, 1, 2).float())
 
         loss_segmentation = sum(losses)
         total_loss = loss_segmentation
